Fixed an issue caused by occasionally returning a reference to a local temporary in the Covert named casts.
Apparently, older versions of clang (and newer versions of gcc) did not
always reliably warn when a reference to a local temporary was being
returned from a function call. Newer versions of clang (i.e., at least >=8.0)
do emit such warnings. This was only a problem within the Covert named
casts, and the solution was to favor reinterpret_cast over static_cast
whenever a local temporary must be created, and a reference to that
local temporary must be returned.